estilizando o input da assinatura e a assinatura modelo
